Controller of Users

+++ backend/src/controllers/UserController.js
@@ -5,6 +5,7 @@ const jwt = require('jsonwebtoken');
 const authConfig = require('../config/auth');
 const fs = require('fs');
 const path = require('path');
+const mailer = require('../modules/mailer');
 
 function hash(password){
 	const saltRounds = 12;
@@ -46,6 +47,18 @@ module.exports = {
 			   latitude,
 			   longitude } = req.body;
 
+		const findEmail = await connection('users').where('email',email).select('email').first();
+
+		const findUser = await connection('users').where('name',name).select('name').first();
+
+		if(findEmail){
+			return res.status(401).json({error:"Email de usuário já cadastrado."});
+		}
+
+		if(findUser){
+			return res.status(401).json({error:"Nome de usuário já cadastrado."});
+		}
+
 		const id = crypto.randomBytes(5).toString('HEX');
 		const password = hash(passwordInput);
 		
@@ -86,11 +99,14 @@ module.exports = {
 		const oldAvatarKey = await connection('uploads').where('user_id', userId).select('key')
 		.first();
 		
-		await fs.unlink(`./temp/uploads/users/${oldAvatarKey.key}`, function(err){
-			if(err) throw err;
-		});
+		if(oldAvatarKey){
+			await fs.unlink(`./temp/uploads/users/${oldAvatarKey.key}`, function(err){
+				if(err) throw err;
+			});
+			
+			await connection('uploads').where('user_id', userId).delete();
+		}
 		
-		await connection('uploads').where('user_id', userId).delete();
 		await connection('users').where('id', userId).delete();
 		return res.send();
 		
@@ -118,6 +134,82 @@ module.exports = {
 			user_id
 		}); 
 		return res.json({sucess:"Imagem carregada com sucesso!" });
+	},
+	recovery: async(req,res) =>{
+		const {email} = req.body;
+		try{
+			const findEmail = await connection('users').where('email',email).select('name')
+			.first();
+
+			if(!findEmail)
+				return res.status(400).json({error: "Usuário não encontrado."});
+
+			const token = crypto.randomBytes(20).toString('HEX');
+
+			const now = new Date();
+			now.setHours(now.getHours()+1);
+
+			await connection('users').where('email',email).update({
+				password_reset_token: token,
+				password_reset_expires: now
+			});
+
+			mailer.sendMail({
+				to: email,
+				from:'etrash@outlook.com.br',
+				template:'auth/forgot_password',
+				context: {token}
+			},(err)=>{
+				if(err){
+					return res.status(400).json({error: "Erro ao enviar o email."})
+				}
+				return res.send();
+			});
+
+		}catch(err){
+			res.status(400).json({error: "Erro ao recuperar a senha, tente de novo."});
+		}
+	},
+	reset: async(req,res)=>{
+		const {email,token,password} = req.body;
+		try{
+			
+			const findEmail = await connection('users').where('email',email).select('email')
+			.first();
+
+			if(!findEmail){
+				return res.status(400).json({error: "Email não encontrado"});
+			}
+
+			const findResetToken = await connection('users').where('email',email).select('password_reset_token')
+			.first();
+
+			if(token !== findResetToken.password_reset_token){
+				return res.status(400).json({error:"Chave para resetar a senha é inválida."});
+			}
+
+			const findResetExpires = await connection('users').where('email',email).select('password_reset_expires')
+			.first();
+
+			const now = new Date();
+
+			if(now > findResetExpires){
+				return res.status(400).json({error:"Sua chave para resetar a senha expirou, pegue uma nova."});
+			}
+
+			const passwordCrypt = hash(password);
+
+			await connection('users').where('email',email).update({
+				password: passwordCrypt,
+				password_reset_token: null,
+				password_reset_expires: null
+			});
+
+			return res.send("Senha resetada com sucesso.");
+
+		}catch(err){
+			return res.status(400).json({error:"Erro ao resetar a senha."});
+		}
 	}
 
 };
